backend/requirements.txt
# Flask and extensions
Flask==3.0.0
flask-pymongo==2.3.0
flask-cors==4.0.0
flask-limiter==3.5.0
flask-socketio==5.3.6
Werkzeug==3.0.1

# WebSocket
python-socketio==5.10.0

# Database
pymongo==4.6.1

# HTTP client
requests==2.31.0
urllib3==2.1.0

# Scheduler
APScheduler==3.10.4

# Geolocation
geoip2==4.8.0
maxminddb==2.5.2

# Security
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Development
pytest==7.4.3
pytest-flask==1.3.0
